+++ include/Neon/Component/NeonLight.h
@@ -13,7 +13,7 @@ namespace Neon
 		Light(const string& name);
 		virtual ~Light();
 
-		virtual void OnUpdate(float now, float timeDelta);
+		virtual void OnUpdate(double now, double timeDelta);
 
 		glm::vec3 position = glm::vec3(0.0f, 10.0f, 0.0f);
 		glm::vec3 direction = glm::vec3(0.0f, -1.0f, 0.0f);
